Change /tree resource
