+++ src/main/java/hive/kirby/controller/HomeController.java
@@ -10,7 +10,6 @@ import org.springframework.web.bind.annotation.RestController;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 
-import static hive.kirby.util.Utils.getTree;
 import static hive.pandora.constant.HiveInternalHeaders.AUTHENTICATED_USER_ID;
 
 @RestController
@@ -26,9 +25,9 @@ public class HomeController {
     final var userRoot = Paths.get(rootDir, userId);
 
     if (!Files.exists(userRoot)) {
-      return new PathNode("root", new PathNode[0]);
+      return PathNode.getEmptyDirInstance("root");
     }
 
-    return new PathNode("root", getTree(userRoot));
+    return new PathNode(userRoot);
   }
 }
